fix: restore build after upstream ISA/Value refactor and make all tests pass
- Restore Runtime struct and fresh_actor_id() counter removed upstream
- Fix CapDown opcode discriminant collision (0x2 -> 0xA2)
- Replace broken low-nibble NaN-boxing with distinct high-16 type tags
- Fix VM operand layouts for Const0/1/2, Not, INeg, Jmp/JmpT/JmpF
- Implement Load/Store/Dup and minimal ArrAlloc/ArrLoad/ArrStore/ArrLen
- Implement Closure opcode (capture-free); CapLoad/CapStore no-op stubs
- Add lexer/parser keywords: and, or, not, then, nil, case
- Add Literal::Nil / Constant::Nil through compiler/typechecker/VM
- Route let-bound lambdas through compile_let_rec for self-reference
- Make match  optional and support  arm keywords
- Update Python marshaling for new Value tagging; implement is_safe_to_marshal
- Fix effect test constant setup
- Update README, SPEC, and PUSH_INSTRUCTIONS for current state
- All 508 unit tests now pass

+++ src/runtime/crdt_manager.rs
@@ -146,7 +146,7 @@ impl CrdtManager {
 
     pub fn create_orset(&mut self) -> (CrdtId, ORSet<String>) {
         let id = CrdtId::new();
-        let set = ORSet::new(self.node_id);
+        let set = ORSet::new(self.node_id as u32);
         self.entries.insert(id, CrdtEntry::ORSet(set.clone()));
         (id, set)
     }
